list: use size_t for position in sheep_list_search()

+++ include/sheep/list.h
@@ -26,7 +26,7 @@ static inline struct sheep_list *sheep_list(sheep_t sheep)
 	return sheep_data(sheep);
 }
 
-int sheep_list_search(struct sheep_list *, sheep_t, unsigned long *);
+int sheep_list_search(struct sheep_list *, sheep_t, size_t *);
 
 void sheep_list_builtins(struct sheep_vm *);
